feat: move StyledInput and StyledSelect components to forms directory; update imports across the application
